backend/requirements.txt
# Web framework
Flask==3.0.3
Werkzeug==3.0.4
gunicorn==23.0.0

# Database
PyMySQL==1.1.1
cryptography==42.0.8

# Auth
PyJWT==2.9.0
bcrypt==4.2.0
Authlib==1.3.2

# Cache / Task queue
redis==5.1.1
celery==5.4.0

# Environment
python-dotenv==1.0.1

# Utilities
Faker==30.8.1

# Testing
pytest==8.3.3
pytest-flask==1.3.0
